🍰 Entire Auth Cycle with User Creation Completed
